Organizando o daily_report
